fixed comments on PR
