+++ front/src/views/index.js
@@ -17,10 +17,10 @@ export { default as MyAccommodation } from '@views/user/Accommodation.vue';
 export { default as MyFood } from '@views/user/Food.vue';
 export { default as MyTransport } from '@views/user/Transport.vue';
 export { default as MyLeisure } from '@views/user/Leisure.vue';
-export { default as RecoveryPassword } from '@views/recovery-password/RecoveryPassword.vue';
 export { default as SignIn } from './sign-in/sign-in.view.vue';
 export { default as SignUp } from './sign-up/sign-up.view.vue';
 export { default as UserCabinet } from '@views/user/UserCabinet.vue';
+export { default as RecoveryPassword } from '@views/recovery-password/RecoveryPassword.vue';
 export {
   default as MyHistoryPreviousTrips
 }
