+++ front/src/views/recovery-password/RecoveryPassword.css
@@ -22,3 +22,19 @@
 .v-stepper {
     box-shadow: none;
 }
+
+.btn-step-1, .btn-step-2 {
+    height: 39px;
+}
+
+.btn-step-1 {
+    width: 159px;
+}
+
+.btn-step-2 {
+    width: 149px;
+}
+
+.input-field {
+    width: 248px;
+}
