+++ front/src/views/recovery-password/RecoveryPassword.html
@@ -36,7 +36,7 @@
               {{$t('recoverypassword.Entuemail')}}
             </div>
             <v-form ref="form" v-model="valid">
-              <div class="input-email" width="248px">
+              <div class="input-email">
                 <mmt-text-input
                   class="input-field"
                   mmt-label="E-mail"
@@ -55,12 +55,10 @@
               <v-row cols="2" justify="center">
                 <v-col justify="center">
                   <v-btn
-                    class="mmt-btn"
+                    class="mmt-btn btn-step-1"
                     rounded
                     color="primary"
                     :disabled="!valid"
-                    height="39px"
-                    width="159px"
                     @click="recoveryPassword"
                     >{{$t('recoverypassword.recover')}}
                   </v-btn>
@@ -70,11 +68,8 @@
                     rounded
                     outlined
                     color="accent lighten-1"
-                    class="mmt-btn"
-                    large
+                    class="mmt-btn btn-step-1"
                     dense
-                    height="39px"
-                    width="159px"
                     @click="$router.push('/');"
                     >{{$t('recoverypassword.back')}}
                   </v-btn>
@@ -92,31 +87,33 @@
             {{$t('recoverypassword.excelentwrk')}}
           </div>
           <v-form ref="form" v-model="valid">
-            <div class="input-firstPass" width="248px">
+            <div class="input-firstPass">
               <mmt-text-input
                 :placeholder="$t('recoverypassword.createpass')"
                 :mmt-label="$t('recoverypassword.password')"
+                class="input-field"
                 rounded
                 filled
                 dense
-                :type="showPass ? 'text' : 'password'"
-                :append-icon="showPass ? 'mdi-eye' : 'mdi-eye-off'"
-                @click:append="showPass = !showPass"
+                :type="showfirstPass ? 'text' : 'password'"
+                :append-icon="showfirstPass ? 'mdi-eye' : 'mdi-eye-off'"
+                @click:append="showfirstPass = !showfirstPass"
                 :rules="[rules.isEmpty, rules.minPassLen]"
                 v-model="firstPass"
                 required
               />
             </div>
-            <div class="input-secondPass" width="248px">
+            <div class="input-secondPass">
               <mmt-text-input
                 :placeholder="$t('recoverypassword.repeatpass')"
                 :mmt-label="$t('recoverypassword.confrimpass')"
+                class="input-field"
                 rounded
                 filled
                 dense
-                :type="showPass ? 'text' : 'password'"
-                :append-icon="showPass ? 'mdi-eye' : 'mdi-eye-off'"
-                @click:append="showPass = !showPass"
+                :type="showsecondPass ? 'text' : 'password'"
+                :append-icon="showsecondPass ? 'mdi-eye' : 'mdi-eye-off'"
+                @click:append="showsecondPass = !showsecondPass"
                 :rules="[rules.isEmpty, rules.minPassLen, rules.samePass]"
                 v-model="secondPass"
                 @input="onInput"
@@ -126,12 +123,10 @@
             <v-row cols="2" justify="center">
               <v-col justify="center">
                 <v-btn
-                  class="mmt-btn"
+                  class="mmt-btn btn-step-2"
                   rounded
                   color="primary"
                   :disabled="!valid"
-                  height="39px"
-                  width="149px"
                   @click="SendPassword"
                   >{{$t('recoverypassword.save')}}
                 </v-btn>
@@ -141,11 +136,8 @@
                   rounded
                   outlined
                   color="accent lighten-1"
-                  class="mmt-btn"
-                  large
+                  class="mmt-btn btn-step-2"
                   dense
-                  height="39px"
-                  width="149px"
                   @click="$router.push('/');"
                   >{{$t('recoverypassword.back')}}
                 </v-btn>
