+++ server/WebAPI/DTO/ForgotPasswordModel.cs
@@ -1,10 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Service.Models
+namespace WebAPI.DTO
 {
-    public class ResetPasswordModel
+    public class ForgotPasswordModel
     {
-        public string Password { get; set; }
+        [Required]
+        [EmailAddress]
         public string Email { get; set; }
     }
 }
