+++ server/WebAPI/DTO/FrontConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace WebAPI.DTO
+{
+	public class FrontConfiguration
+	{
+		public string AddressFront { get; set; }
+	}
+}
