+++ server/WebAPI/Services/AuthService.cs
@@ -8,33 +8,33 @@ using Microsoft.IdentityModel.Tokens;
 
 namespace WebAPI.Services
 {
-    public class AuthService
-    {
-        private readonly IConfiguration _configuration;
-        
-        public AuthService(IConfiguration configuration)
-        {
-            _configuration = configuration;
-        }
+  public class AuthService
+  {
+    private readonly IConfiguration _configuration;
 
-        public string MakeToken(User user, bool oneDay = true)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_configuration["Jwt:Key"]);
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(new[] 
-                {
+    public AuthService(IConfiguration configuration)
+    {
+      _configuration = configuration;
+    }
+
+    public string MakeToken(User user, bool oneDay = true)
+    {
+      var tokenHandler = new JwtSecurityTokenHandler();
+      var key = Encoding.ASCII.GetBytes(_configuration["Jwt:Key"]);
+      var tokenDescriptor = new SecurityTokenDescriptor
+      {
+        Subject = new ClaimsIdentity(new[]
+          {
                     new Claim(ClaimTypes.Name, user.Id.ToString()),
                     new Claim(ClaimTypes.Email, user.Email)
                 }),
-                Expires = oneDay ? DateTime.UtcNow.AddDays(1) : DateTime.UtcNow.AddMinutes(10),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
-                Issuer = _configuration["Jwt:Issuer"],
-                Audience = _configuration["Jwt:Audience"]
-            };
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
-        }
+        Expires = oneDay ? DateTime.UtcNow.AddDays(1) : DateTime.UtcNow.AddMinutes(10),
+        SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
+        Issuer = _configuration["Jwt:Issuer"],
+        Audience = _configuration["Jwt:Audience"]
+      };
+      var token = tokenHandler.CreateToken(tokenDescriptor);
+      return tokenHandler.WriteToken(token);
     }
+  }
 }
