+++ server/WebAPI/Services/UserService.cs
@@ -1,73 +1,70 @@
 using Entities;
 using Entities.Models;
-using System;
-using System.Security.Cryptography;
-using System.Text;
 
 namespace WebAPI.Services
 {
-    public class UserService
+  public class UserService
+  {
+
+    private readonly UserRepository _userRepository;
+
+    public UserService(UserRepository userRepository)
     {
-
-        private readonly UserRepository _userRepository;
-        
-        public UserService(UserRepository userRepository)
-        {
-            _userRepository = userRepository;
-        }
-
-        public User GetUser(string email, string password)
-        {
-            var user = _userRepository.FindUserByEmail(email);
-            if (user != null && CryptoUtils.VerifyPassword(password, user.Password))
-            {
-                return user;
-            }
-            return null;
-        }
-
-        public bool IsUserExist(string email)
-        {
-            return _userRepository.FindUserByEmail(email) != null;
-        }
-
-        public void CreateUser(string email, string password)
-        {
-            var user = new User
-            {
-                Email = email,
-                Password = CryptoUtils.HashPassword(password)
-            };
-            
-            _userRepository.CreateUser(user);
-        }
-        public User GetUser(string email)
-        {
-            return _userRepository.FindUserByEmail(email);
-        }
-
-        public bool UpdateUserPassword(string email, string newPassword)
-        {
-            try
-            {
-                var user = GetUser(email);
-                if(user != null)
-                {
-                    user.Password = CryptoUtils.HashPassword(newPassword);
-                    _userRepository.UpdateUserPassword(user);
-                    return true;
-                }
-                else
-                {
-                    CreateUser(email, newPassword);
-                    return true;
-                }
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
+      _userRepository = userRepository;
     }
+
+    public User GetUser(string email, string password)
+    {
+      var user = _userRepository.FindUserByEmail(email);
+      if(user != null && CryptoUtils.VerifyPassword(password, user.Password))
+      {
+        return user;
+      }
+      return null;
+    }
+
+    public bool IsUserExist(string email)
+    {
+      return _userRepository.FindUserByEmail(email) != null;
+    }
+
+    public void CreateUser(string email, string password)
+    {
+      var user = new User
+      {
+        Email = email,
+        Password = CryptoUtils.HashPassword(password)
+      };
+
+      _userRepository.CreateUser(user);
+    }
+    public User GetUser(string email)
+    {
+      return _userRepository.FindUserByEmail(email);
+    }
+
+    public bool UpdateUserPassword(string email, string newPassword)
+    {
+      try
+      {
+        var user = GetUser(email);
+        if(user != null)
+        {
+          user.Password = CryptoUtils.HashPassword(newPassword);
+          _userRepository.UpdateUserPassword(user);
+          return true;
+        }
+        else
+        {
+          CreateUser(email, newPassword);
+          return true;
+        }
+      }
+      catch
+      {
+        return false;
+      }
+    }
+
+  }
 }
