+++ server/WebAPI/Startup.cs
@@ -9,119 +9,121 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
-using Service.Interfaces;
-using Service.MailService;
-using Service.Models;
 using Microsoft.OpenApi.Models;
 using WebAPI.Options;
 using WebAPI.DTO;
 using WebAPI.Services;
+using WebAPI.Interfaces;
 
 namespace WebAPI
 {
-    public class Startup
+  public class Startup
+  {
+    readonly string VueCorsPolicy = "_vueCorsPolicy";
+    public Startup(IConfiguration configuration)
     {
-        readonly string VueCorsPolicy = "_vueCorsPolicy";
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddCors(options =>
-            {
-                options.AddPolicy(name: VueCorsPolicy,
-                                  builder =>
-                                  {
-                                      builder
-                                        .AllowAnyHeader()
-                                        .AllowAnyMethod()
-                                        .AllowCredentials()
-                                        .WithOrigins("http://localhost:8080");
-                                  });
-            });
-
-            services.AddDbContext<RepositoryContext>(options => 
-                options.UseSqlServer(Configuration.GetConnectionString("mssqlConnection")));
-            services.AddSingleton(Configuration);
-            services.AddScoped<UserRepository>();
-            services.AddScoped<UserService>();
-            services.AddSingleton<AuthService>();
-            services.AddScoped<TripRepository>();
-            services.AddScoped<TripService>();
-            services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-                .AddJwtBearer(x =>
-            {
-                x.RequireHttpsMetadata = false;
-                x.SaveToken = true;
-                x.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(Configuration["Jwt:Key"])),
-                    ValidateIssuer = true,
-                    ValidIssuer = Configuration["Jwt:Issuer"],
-                    ValidateAudience = true,
-                    ValidAudience = Configuration["Jwt:Audience"]
-                };
-            });
-			var emailConfig = Configuration.GetSection("EmailConfiguration").Get<EmailConfiguration>();
-			services.AddSingleton(emailConfig);
-			services.AddScoped<IEmailSender, EmailSender>();
-
-			services.AddControllers();
-
-            services.AddSwaggerGen(x =>
-            {
-                x.SwaggerDoc("v1", new OpenApiInfo { Title = "MyTrip Api", Version = "v1" });
-            });
-        
-            var mappingConfig = new MapperConfiguration(mc =>
-            {
-              mc.AddProfile(new MappingProfile());
-            });
-
-            services.AddSingleton(mappingConfig.CreateMapper());
-        }
-    
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            var swaggerOptions = new SwaggerOptions();
-            Configuration.GetSection(nameof(SwaggerOptions)).Bind(swaggerOptions);
-
-            app.UseSwagger(option =>
-            {
-                option.RouteTemplate = swaggerOptions.JsonRoute;
-            });
-
-            app.UseSwaggerUI(option =>
-            {
-                option.SwaggerEndpoint(swaggerOptions.UiEndpoint, swaggerOptions.Description);
-            });
-
-            app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseCors(VueCorsPolicy);
-
-            app.UseAuthentication();
-            
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });      
-        }
+      Configuration = configuration;
     }
+
+    public IConfiguration Configuration { get; }
+
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
+    {
+
+      var frontConfiguration = Configuration.GetSection("Front").Get<FrontConfiguration>();
+      services.AddCors(options =>
+      {
+        options.AddPolicy(name: VueCorsPolicy,
+                                builder =>
+                                {
+                              builder
+                                      .AllowAnyHeader()
+                                      .AllowAnyMethod()
+                                      .AllowCredentials()
+                                      .WithOrigins(frontConfiguration.AddressFront);
+                            });
+      });
+
+      services.AddSingleton(frontConfiguration);
+      services.AddScoped<ResetPasswordService>();
+      services.AddDbContext<RepositoryContext>(options =>
+          options.UseSqlServer(Configuration.GetConnectionString("mssqlConnection")));
+      services.AddSingleton(Configuration);
+      services.AddScoped<UserRepository>();
+      services.AddScoped<UserService>();
+      services.AddSingleton<AuthService>();
+      services.AddScoped<TripRepository>();
+      services.AddScoped<TripService>();
+      services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+          .AddJwtBearer(x =>
+      {
+        x.RequireHttpsMetadata = false;
+        x.SaveToken = true;
+        x.TokenValidationParameters = new TokenValidationParameters
+        {
+          ValidateIssuerSigningKey = true,
+          IssuerSigningKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(Configuration["Jwt:Key"])),
+          ValidateIssuer = true,
+          ValidIssuer = Configuration["Jwt:Issuer"],
+          ValidateAudience = true,
+          ValidAudience = Configuration["Jwt:Audience"]
+        };
+      });
+      var emailConfig = Configuration.GetSection("EmailConfiguration").Get<EmailConfiguration>();
+      services.AddSingleton(emailConfig);
+      services.AddScoped<IEmailSender, EmailSender>();
+
+      services.AddControllers();
+
+      services.AddSwaggerGen(x =>
+      {
+        x.SwaggerDoc("v1", new OpenApiInfo { Title = "MyTrip Api", Version = "v1" });
+      });
+
+      var mappingConfig = new MapperConfiguration(mc =>
+      {
+        mc.AddProfile(new MappingProfile());
+      });
+
+      services.AddSingleton(mappingConfig.CreateMapper());
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+      if(env.IsDevelopment())
+      {
+        app.UseDeveloperExceptionPage();
+      }
+
+      var swaggerOptions = new SwaggerOptions();
+      Configuration.GetSection(nameof(SwaggerOptions)).Bind(swaggerOptions);
+
+      app.UseSwagger(option =>
+      {
+        option.RouteTemplate = swaggerOptions.JsonRoute;
+      });
+
+      app.UseSwaggerUI(option =>
+      {
+        option.SwaggerEndpoint(swaggerOptions.UiEndpoint, swaggerOptions.Description);
+      });
+
+      app.UseHttpsRedirection();
+
+      app.UseRouting();
+
+      app.UseCors(VueCorsPolicy);
+
+      app.UseAuthentication();
+
+      app.UseAuthorization();
+
+      app.UseEndpoints(endpoints =>
+      {
+        endpoints.MapControllers();
+      });
+    }
+  }
 }
