+++ server/WebAPI/appsettings.json
@@ -18,5 +18,8 @@
     "Port": 587,
     "Username": "makemetripapp@gmail.com",
     "Password": "your password"
+  },
+  "Front": {
+    "AddressFront": "http://localhost:8081"
   }
 }
